+++ src/pages/storage/Storage.tsx
@@ -1,25 +1,21 @@
 import styled from "@emotion/styled";
 import Title from "@/shared/component/Title";
-import defaultProfile from "@/assets/images/defaultProfile.svg";
-import Button from "@/shared/component/Button";
 import { useState } from "react";
-import Like from "@/assets/images/like.svg";
-import Comment from "@/assets/images/comment.svg";
 import { useLocation, useNavigate, useParams } from "react-router-dom";
 import useOtherUser from "@/api/useOtherUser";
 import useFollowCount from "@/api/useFollowCount";
 import Loading from "@/shared/component/Loading";
 import { useUserStore } from "@/stores/userStore";
-import useFollowStatus from "@/pages/followInfo/hooks/useFollowStatus";
+import useFollowStatus from "@/api/useFollowStatus";
 import useLikedPlaylists from "@/api/useLikedPlaylists";
 import useMyPlaylists from "@/api/useMyPlaylists";
-import backgroundImage from "@/assets/images/backGround.png";
-import Dropbox from "@/shared/component/Dropbox";
 import Modal from "@/shared/component/Modal";
 import { softDeletePlaylist } from "@/db/playlist";
 import { useQueryClient, useQuery } from "@tanstack/react-query";
 import { getMyLikedPlaylistIds } from "@/db/like";
-import { ReactSVG } from "react-svg";
+import StoragePlaylistCard from "@/pages/storage/component/StoragePlaylistCard";
+import StorageProfileCard from "@/pages/storage/component/StorageProfileCard";
+import { TabMenu, TabButton } from "@/shared/component/Tab";
 
 const Storage = () => {
   const [activeTab, setActiveTab] = useState<"left" | "right">("left");
@@ -63,11 +59,6 @@ const Storage = () => {
     isUnfollowPending,
   } = useFollowStatus(targetId);
 
-  const getProfileImageUrl = (userImg?: string | null) => {
-    if (!userImg) return defaultProfile;
-    return userImg;
-  };
-
   const handleNavigate = (tab: "follower" | "following") => {
     navigate(`/storage/${randomId}/follow-info?tab=${tab}`);
   };
@@ -95,88 +86,35 @@ const Storage = () => {
 
       <StorageWrapper>
         <FixedHeaderArea>
-          <ProfileWrapper>
-            <ProfileCardTop>
-              <ImageArea src={getProfileImageUrl(userData?.user_img)} />
-              <ProfileInfo>
-                <NickName>{userData?.nickname}</NickName>
-                <InfoItemWrapper>
-                  <InfoItem onClick={() => handleNavigate("follower")}>
-                    <InfoCount>{followerCount}</InfoCount>
-                    <InfoLabel>팔로워</InfoLabel>
-                  </InfoItem>
-                  <InfoItem onClick={() => handleNavigate("following")}>
-                    <InfoCount>{followingCount}</InfoCount>
-                    <InfoLabel>팔로잉</InfoLabel>
-                  </InfoItem>
-                  <InfoItem>
-                    <InfoCount>{myPlaylists.length}</InfoCount>
-                    <InfoLabel>리스트</InfoLabel>
-                  </InfoItem>
-                </InfoItemWrapper>
-                {isMyPage ? (
-                  !isProfilePage && (
-                    <Button size="small" onClick={() => navigate("/profile")}>
-                      프로필 수정
-                    </Button>
-                  )
-                ) : (
-                  <Button
-                    size="small"
-                    btnColor={isFollowing ? "white" : "pink"}
-                    onClick={() => {
-                      if (!currentUser?.random_id || !targetId) return;
-                      if (isFollowing) {
-                        handleUnfollow();
-                      } else {
-                        handleFollow();
-                      }
-                    }}
-                    disabled={
-                      isFollowPending ||
-                      isUnfollowPending ||
-                      !currentUser?.random_id ||
-                      !targetId
-                    }
-                  >
-                    {isFollowPending || isUnfollowPending
-                      ? "처리 중..."
-                      : isFollowing
-                        ? "팔로잉"
-                        : "팔로우"}
-                  </Button>
-                )}
-              </ProfileInfo>
-            </ProfileCardTop>
-            <ProfileCardBottom>
-              <BioArea>{userData?.sort_intro}</BioArea>
-              <HashTagArea>
-                {userData?.artist_hash_tag?.split(" ").map((word, i) =>
-                  word.startsWith("#") ? (
-                    <span key={i} className="hashtag">
-                      {word}
-                    </span>
-                  ) : (
-                    <span key={i}> {word} </span>
-                  ),
-                )}
-              </HashTagArea>
-            </ProfileCardBottom>
-          </ProfileWrapper>
+          <StorageProfileCard
+            userData={userData!}
+            isMyPage={isMyPage}
+            isProfilePage={isProfilePage}
+            followerCount={followerCount}
+            followingCount={followingCount}
+            playlistCount={myPlaylists.length}
+            isFollowing={isFollowing}
+            isFollowPending={isFollowPending}
+            isUnfollowPending={isUnfollowPending}
+            onFollow={handleFollow}
+            onUnfollow={handleUnfollow}
+            onNavigateToProfile={() => navigate("/profile")}
+            onNavigateToFollowInfo={handleNavigate}
+          />
 
           <TabMenu>
-            <TabLeft
+            <TabButton
               isActive={activeTab === "left"}
               onClick={() => setActiveTab("left")}
             >
               리스트
-            </TabLeft>
-            <TabRight
+            </TabButton>
+            <TabButton
               isActive={activeTab === "right"}
               onClick={() => setActiveTab("right")}
             >
               하트
-            </TabRight>
+            </TabButton>
           </TabMenu>
         </FixedHeaderArea>
 
@@ -186,49 +124,16 @@ const Storage = () => {
               const isLiked = likedIds.includes(item.p_id);
 
               return (
-                <VideoWrapper
+                <StoragePlaylistCard
                   key={item.p_id}
-                  onClick={() => navigate(`/playlist/${item.p_id}`)}
-                >
-                  <VideoArea src={item.cover_img_path || backgroundImage} />
-                  <Meta>
-                    <DetailArea>
-                      <Count>동영상 {item.video_count}개</Count>
-                      <IconGroup>
-                        <span className="like">
-                          <ReactSVG
-                            src={Like}
-                            wrapper="span"
-                            className={`likeSvg ${isLiked ? "active" : "inactive"}`}
-                          />
-                          {item.like_count}
-                        </span>
-
-                        <span className="Comment">
-                          <img src={Comment} alt="댓글" />
-                          {item.comment_count}
-                        </span>
-                        {isMyPage &&
-                          activeTab === "left" &&
-                          item.random_id === currentUser?.random_id && (
-                            <div
-                              onClick={(e) => {
-                                e.stopPropagation();
-                              }}
-                            >
-                              <Dropbox
-                                variant="icon"
-                                onChange={(action) =>
-                                  handleIconAction(action, item.p_id)
-                                }
-                              />
-                            </div>
-                          )}
-                      </IconGroup>
-                    </DetailArea>
-                    <VideoTitle>{item.playlist_title}</VideoTitle>
-                  </Meta>
-                </VideoWrapper>
+                  item={item}
+                  isLiked={isLiked}
+                  isMyPage={isMyPage}
+                  activeTab={activeTab}
+                  currentUserId={currentUser?.random_id}
+                  onNavigate={(id) => navigate(`/playlist/${id}`)}
+                  onIconAction={handleIconAction}
+                />
               );
             },
           )}
@@ -270,154 +175,6 @@ const Storage = () => {
 
 export default Storage;
 
-const ProfileWrapper = styled.div`
-  width: 600px;
-  height: 280px;
-  padding: 0px 40px;
-`;
-
-const ProfileCardTop = styled.div`
-  display: flex;
-  padding: 18px 20px;
-  gap: 100px;
-`;
-
-const ImageArea = styled.img`
-  width: 165px;
-  height: 165px;
-  border-radius: 50%;
-  object-fit: cover;
-`;
-
-const ProfileInfo = styled.div`
-  width: 200px;
-  height: 165px;
-  display: flex;
-  flex-direction: column;
-  justify-content: space-between;
-`;
-
-const NickName = styled.span`
-  font-size: var(--font-size-subtitle);
-  font-weight: 700;
-`;
-
-const InfoItemWrapper = styled.div`
-  display: flex;
-  justify-content: space-between;
-  gap: 25px;
-`;
-
-const InfoItem = styled.div`
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  gap: 7px;
-  cursor: pointer;
-`;
-
-const InfoCount = styled.span`
-  font-size: var(--font-size-large);
-  font-weight: 700;
-`;
-
-const InfoLabel = styled.span`
-  font-size: var(--font-size-primary);
-  font-weight: 400;
-`;
-
-const ProfileCardBottom = styled.div`
-  margin-top: 10px;
-  margin-left: 8px;
-  display: flex;
-  flex-direction: column;
-  gap: 5px;
-`;
-
-const BioArea = styled.span`
-  font-size: var(--font-size-large);
-  color: var(--text-primary);
-  font-weight: 400;
-  line-height: 18px;
-`;
-
-const HashTagArea = styled.div`
-  display: flex;
-  flex-wrap: wrap;
-  align-items: center; // 🔥 세로 중앙 정렬
-  gap: 8px;
-  margin-top: 4px;
-
-  .hashtag {
-    display: inline-flex;
-    align-items: center;
-    justify-content: center;
-    background-color: var(--primary);
-    color: white;
-    padding: 6px 12px;
-    border-radius: 999px;
-    font-size: var(--font-size-primary);
-    font-weight: 500;
-    line-height: 1; // 🔥 줄간격 최소화로 위아래 여백 균등
-  }
-
-  span:not(.hashtag) {
-    font-size: var(--font-size-primary);
-    color: var(--text-primary);
-    display: inline-flex;
-    align-items: center; // 🔥 비해시태그 텍스트도 정렬 맞춤
-    line-height: 1;
-  }
-`;
-
-const TabMenu = styled.div`
-  width: 600px;
-  height: 60px;
-  display: flex;
-  font-size: var(--font-size-large);
-`;
-
-const TabLeft = styled.div<{ isActive: boolean }>`
-  width: 300px;
-  height: 60px;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  cursor: pointer;
-  font-weight: ${(props) => (props.isActive ? 700 : 400)};
-  border-bottom: ${(props) =>
-    props.isActive
-      ? `3px solid var(--primary)`
-      : `3px solid var(--disabled-2)`};
-  color: ${(props) =>
-    props.isActive ? `var(--primary)` : `var(--text-primary)`};
-  &:hover {
-    background-color: var(--profile-background);
-    transition: background-color 0.3s ease-in-out;
-  }
-`;
-
-const TabRight = styled.div<{ isActive: boolean }>`
-  width: 300px;
-  height: 60px;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  cursor: pointer;
-  font-weight: ${(props) => (props.isActive ? 700 : 400)};
-  border-bottom: ${(props) =>
-    props.isActive
-      ? `3px solid var(--primary)`
-      : `3px solid var(--disabled-2)`};
-  color: ${(props) =>
-    props.isActive ? `var(--primary)` : `var(--text-primary)`};
-  &:hover {
-    background-color: var(--profile-background);
-    transition: background-color 0.3s ease-in-out;
-  }
-`;
-
 const StorageWrapper = styled.div`
   display: flex;
   flex-direction: column;
@@ -443,98 +200,6 @@ const ScrollablePlaylistArea = styled.div`
   &::after {
     content: "";
     flex-basis: 100%;
-    height: 120px; // ✅ 이게 하단 여백 역할!
+    height: 120px;
   }
-`;
-
-// const PlaylistArea = styled.div`
-//   display: flex;
-//   gap: 15px 0;
-//   padding: 15px 40px 85px 40px;
-//   overflow-y: auto;
-//   flex-wrap: wrap;
-//   max-height: 800px;
-//   justify-content: space-between;
-//   align-items: flex-start;
-// `;
-
-const VideoWrapper = styled.div`
-  width: 250px;
-  height: 220px;
-  border-radius: 15px;
-  box-shadow: 0px 0px 5px 0px rgba(0, 0, 0, 0.2);
-  cursor: pointer;
-  object-fit: cover;
-
-  &:hover {
-    background-color: var(--profile-background);
-    transition: background-color 0.3s ease-in-out;
-    box-shadow: 0px 0px 10px 0px rgba(0, 0, 0, 0.2);
-  }
-`;
-
-const VideoArea = styled.img`
-  background-color: var(--text-secondary);
-  width: 224px;
-  height: 126px;
-  border-radius: 5px;
-  margin: 11px 13px 7px 13px;
-`;
-
-const Meta = styled.div`
-  display: flex;
-  justify-content: space-between;
-  font-size: var(--font-size-small);
-  color: var(--text-secondary);
-  gap: 9px;
-  flex-direction: column;
-  margin: 0 13px 13px 13px;
-`;
-
-const DetailArea = styled.div`
-  display: flex;
-  flex-direction: row;
-  justify-content: space-between;
-`;
-
-const Count = styled.span``;
-
-const IconGroup = styled.div`
-  display: flex;
-  gap: 10px;
-  span {
-    display: flex;
-    align-items: center;
-    gap: 4px;
-    img {
-      width: 14px;
-      height: 14px;
-    }
-    .likeSvg svg {
-      width: 14px;
-      height: 14px;
-      display: block;
-      color: var(--text-secondary); /* 기본 회색 */
-    }
-
-    .likeSvg.active svg {
-      color: var(--primary); /* 좋아요 눌렀을 때 */
-    }
-
-    .likeSvg.inactive svg {
-      stroke: var(--text-secondary); /* 비활성 테두리 */
-      fill: none; /* 비활성 내부는 비워둠 */
-    }
-  }
-`;
-
-const VideoTitle = styled.div`
-  font-size: var(--font-size-primary);
-  font-weight: 700;
-  color: var(--text-primary);
-  display: -webkit-box;
-  -webkit-line-clamp: 2;
-  -webkit-box-orient: vertical;
-  overflow: hidden;
-  text-overflow: ellipsis;
 `;
